refactor: 관리자 페이지 대시보드 리스트 DB 연동 및 컴포넌트 분리
기존 page.tsx를 서버 컴포넌트로 전환해 DB에서 브랜드 및 차량 정보 조회,
UI 및 상태 관리 로직 분리, 하드코딩 데이터 제거
