+++ src/app/admin/AdminClient.tsx
@@ -1,0 +1,200 @@
+"use client";
+
+import styles from './MainHome.module.scss';
+import Image from "next/image";
+import {Card, CardContent, Drawer, Button, Modal, Box, IconButton, TextField} from "@mui/material";
+import Icon from "@mdi/react"
+import {mdiClose} from "@mdi/js";
+import {useState, useEffect} from "react";
+import {useRouter} from "next/navigation";
+import {createBrand} from "@/app/admin/actions";
+
+import AvanteNSide from "./assets/avanten_side.png";
+
+interface CarModel{
+    id: number;
+    car_name: string;
+    price: number;
+    liter_size: number;
+    fuel_efficiency: number;
+    car_image?: any;
+}
+
+interface CarBrand{
+    brand_id: number;
+    brand_name: string;
+    CarModel: CarModel[];
+}
+
+interface CountryGroup{
+    brand_country: string;
+    CarBrand: CarBrand[];
+}
+
+interface AdminClientProps{
+    initialData: CountryGroup[];
+}
+
+export default function AdminClient({initialData}: AdminClientProps){
+    const router = useRouter();
+
+    const firstBrandId = initialData[0]?.CarBrand[0]?.brand_id || 0;
+    const [activeBrandId, setActiveBrandId] = useState<number>(firstBrandId);
+
+    const [brandName, setBrandName] = useState("");
+    const [brandCountry, setBrandCountry] = useState("");
+
+    const [openAddBrandModal, setOpenAddBrandModal] = useState(false);
+    const handleOpenAddBrandModal = () => setOpenAddBrandModal(true);
+    const handleCloseAddBrandModal = () => setOpenAddBrandModal(false);
+
+    const [openEditBrandModal, setOpenEditBrandModal] = useState(false);
+    const handleOpenEditBrandModal = () => setOpenEditBrandModal(true);
+    const handleCloseEditBrandModal = () => setOpenEditBrandModal(false);
+
+    const [openAddCarModal, setOpenAddCarModal] = useState(false);
+    const handleOpenAddCarModal = () => setOpenAddCarModal(true);
+    const handleCloseAddCarModal = () => setOpenAddCarModal(false);
+
+    const handleSaveBrand = async ()=>{
+        const result = await createBrand(brandName, brandCountry);
+
+        if (result.success){
+            alert(result.message);
+            setBrandName("");
+            setBrandCountry("");
+            handleCloseAddBrandModal();
+            router.refresh();
+        } else {
+            alert(result.message);
+        }
+    };
+
+    return (
+        <div className={styles.MainHomeStyle}>
+            <div className={styles.MainHomeContainer}>
+                <div className={styles.MainHomeContent}>
+                    <Card className={styles.MainHomeCardStyle}>
+                        <Drawer variant="permanent" anchor="left" PaperProps={{ style: {position: "absolute", height: "100%", width: "180px", overflowY: "auto", border: "none", boxShadow: "none", }}}>
+                            <span className={styles.MainHomeLeftContentTitle}>대시보드</span>
+                            <div className={styles.MainHomeLeftContentItemContainer}>
+                                {initialData.flatMap(country => country.CarBrand.map(brand => (
+                                        <span
+                                            key={brand.brand_id}
+                                            className={`${styles.MainHomeLeftContentItem} ${activeBrandId === brand.brand_id ? styles.active : ''}`}
+                                            onClick={() => setActiveBrandId(brand.brand_id)}
+                                        >
+                                            {brand.brand_name}
+                                        </span>
+                                    ))
+                                )}
+                            </div>
+                            <Button variant="contained" className={styles.MainHomeLeftContentBottomButton} onClick={handleOpenAddBrandModal}>브랜드 추가</Button>
+                        </Drawer>
+                        <CardContent style={{ height: '100%', overflowY: 'auto', marginLeft: '180px'}}>
+                            <div>
+                                {initialData.flatMap(country =>
+                                    country.CarBrand
+                                        .filter(brand => activeBrandId === null || activeBrandId === brand.brand_id)
+                                        .map(brand => (
+                                            <div className={styles.MainHomeCardTitlesContainer} key={brand.brand_id}>
+                                                <span className={styles.MainHomeCardMainTitle}>{brand.brand_name}</span>
+                                                <span className={styles.MainHomeCardSubTitle}>국가: {country.brand_country}</span>
+                                            </div>
+                                        ))
+                                )}
+                            </div>
+                            <div className={styles.MainHomeCardDividerContainer}>
+                                <span className={styles.MainHomeCardDividerTitle}>차량 목록</span>
+                                <div className={styles.MainHomeCardDivider}/>
+                            </div>
+                            <div className={styles.MainHomeCardCarList}>
+                                {initialData.flatMap(country =>
+                                    country.CarBrand.flatMap(brand =>
+                                        brand.CarModel?.filter(() =>
+                                            activeBrandId === null || activeBrandId === brand.brand_id
+                                        )?.map(car => (
+                                            <div key={car.id} className={styles.MainHomeCardCarInfoStyle}>
+                                                <Image
+                                                    className={styles.MainHomeCardCarInfoCarImage}
+                                                    src={car.car_image || AvanteNSide}
+                                                    alt={car.car_name}
+                                                />
+                                                <div className={styles.MainHomeCardCarInfoContainer}>
+                                                    <div className={styles.MainHomeCardCarInfoTitleContent}>
+                                                        <span className={styles.MainHomeCardCarInfoCarName}>{car.car_name}</span>
+                                                    </div>
+                                                    <span className={styles.MainHomeCardCarInfoCarPrice}>{car.price}만원 ~</span>
+                                                    <span className={styles.MainHomeCardCarInfoCarSpec}>배기량: ~ {car.liter_size}cc</span>
+                                                    <span className={styles.MainHomeCardCarInfoCarSpec}>연비: ~ {car.fuel_efficiency} km/l</span>
+                                                </div>
+                                                <div className={styles.MainHomeCardCarInfoSelectButtonGroupStyle}>
+                                                    <Button className={styles.MainHomeCardCarInfoSelectButton} onClick={() => router.push("/admin/EditCar")}>수정하기</Button>
+                                                    <Button className={styles.MainHomeCardCarInfoSelectButton}>삭제하기</Button>
+                                                </div>
+                                            </div>
+                                        ))
+                                    )
+                                )}
+                                <div className={styles.MainHomeCardCarAddButtonWrapper}>
+                                    <Button className={styles.MainHomeCardBrandDeleteButton}>브랜드 삭제</Button>
+                                    <Button className={styles.MainHomeCardBrandEditButton} onClick={handleOpenEditBrandModal}>브랜드 정보 수정</Button>
+                                    <Button className={styles.MainHomeCardCarAddButton} onClick={handleOpenAddCarModal}>차량 추가하기</Button>
+                                </div>
+                            </div>
+                        </CardContent>
+                    </Card>
+                </div>
+            </div>
+            <Modal open={openAddBrandModal} onClose={handleCloseAddBrandModal}>
+                <Box className={styles.MainHomeModalBoxStyle}>
+                    <div className={styles.MainHomeModalBoxTitleWrapper}>
+                        <span className={styles.MainHomeModalBoxMainTitle}>브랜드 추가</span>
+                        <IconButton onClick={handleCloseAddBrandModal}>
+                            <Icon path={mdiClose} size={1} />
+                        </IconButton>
+                    </div>
+                    <div className={styles.MainHomeModalBoxTextFieldWrapper}>
+                        <input className={styles.MainHomeModalBoxTextField} value={brandName} onChange={(e) => setBrandName(e.target.value)} placeholder="브랜드 명을 입력해주세요."/>
+                        <input className={styles.MainHomeModalBoxTextField} value={brandCountry} onChange={(e) => setBrandCountry(e.target.value)} placeholder="브랜드 국가를 입력해주세요."/>
+                        <Button className={styles.MainHomeModalBoxSaveButton} variant="contained" onClick={handleSaveBrand}>저장하기</Button>
+                    </div>
+                </Box>
+            </Modal>
+            <Modal open={openEditBrandModal} onClose={handleCloseEditBrandModal}>
+                <Box className={styles.MainHomeModalBoxStyle}>
+                    <div className={styles.MainHomeModalBoxTitleWrapper}>
+                        <span className={styles.MainHomeModalBoxMainTitle}>브랜드 정보 수정</span>
+                        <IconButton onClick={handleCloseEditBrandModal}>
+                            <Icon path={mdiClose} size={1} />
+                        </IconButton>
+                    </div>
+                    <div className={styles.MainHomeModalBoxTextFieldWrapper}>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="변경할 브랜드 명을 입력해주세요."/>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="변경할 브랜드 국가를 입력해주세요."/>
+                        <Button className={styles.MainHomeModalBoxSaveButton} variant="contained">수정하기</Button>
+                    </div>
+                </Box>
+            </Modal>
+            <Modal open={openAddCarModal} onClose={handleCloseAddCarModal}>
+                <Box className={styles.MainHomeModalBoxStyle}>
+                    <div className={styles.MainHomeModalBoxTitleWrapper}>
+                        <span className={styles.MainHomeModalBoxMainTitle}>차량 추가</span>
+                        <IconButton onClick={handleCloseAddCarModal}>
+                            <Icon path={mdiClose} size={1} />
+                        </IconButton>
+                    </div>
+                    <div className={styles.MainHomeModalBoxTextFieldWrapper}>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="차량명을 입력해주세요."/>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="시작 가격을 입력해주세요."/>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="배기량을 정수로 입력해주세요."/>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="연비를 소수점 한 자리까지 입력해주세요."/>
+                        <input className={styles.MainHomeModalBoxTextField} placeholder="바디 타입을 입력해주세요."/>
+                        <Button className={styles.MainHomeModalBoxSaveButton} variant="contained">추가하기</Button>
+                    </div>
+                </Box>
+            </Modal>
+        </div>
+    );
+}
+
